+++ api/src/main/scala/json2struct/api/Server.scala
@@ -4,11 +4,15 @@ import akka.actor.ActorSystem
 import akka.http.scaladsl.Http
 import akka.http.scaladsl.model.StatusCodes
 import akka.http.scaladsl.server.Directives
+import com.typesafe.config.ConfigFactory
 import json2struct.Conf.APP_CONF
 import json2struct.Converter
 import json2struct.Printer.Syntax.toPrinterOps
 import json2struct.api.Conf.HttpPort
-import json2struct.api.JsonSupport.Json
+import json2struct.api.JsonSupport.*
+import spray.json.*
+
+import scala.jdk.CollectionConverters.MapHasAsJava
 
 object Server extends Directives with JsonSupport {
 
@@ -19,9 +23,9 @@ object Server extends Directives with JsonSupport {
       pathPrefix("v1" / "convert") {
         path("json") {
           post {
-            entity(as[Json]) { json =>
+            entity(as[JsonBody]) { body =>
               complete {
-                Converter.convertJson(json.json, json.name)
+                Converter.convertJson(body.json, body.name)
                   .map(_.print())
                   .mkString(System.lineSeparator())
               }
@@ -51,7 +55,25 @@ object Server extends Directives with JsonSupport {
               }
             }
           }
-        }
+        } ~
+          path("struct") {
+            post {
+              optionalHeaderValueByName("config") { config =>
+                entity(as[String]) { struct =>
+                  val conf = config.fold[Map[String, Any]](Map.empty) { conf =>
+                    conf.parseJson.convertTo[Map[String, Any]]
+                  }.asJava
+                  complete {
+                    Converter.convertStruct(
+                        struct,
+                        ConfigFactory.parseMap(conf).withFallback(APP_CONF))
+                      .map(_.print())
+                      .mkString(System.lineSeparator())
+                  }
+                }
+              }
+            }
+          }
       },
 
       path("health") {
@@ -64,4 +86,5 @@ object Server extends Directives with JsonSupport {
     Http().newServerAt("0.0.0.0", HttpPort).bind(route)
     println(s"Server started on 0.0.0.0:$HttpPort")
   }
+
 }