+++ CHANGELOG.md
@@ -24,3 +24,4 @@
 * add snake-case conf to support converting to snake-case style json props
 * add a new json2struct endpoint with name as query parameter and json as body
 * support to release latest docker image based on newest code in github ci
+* add a new struct2json endpoint to support converter customization with config header